backend/app/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Konstantin Freidlin <freidlinks@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.87.0"}
uvicorn = {extras = ["standart"], version = "^0.19.0"}
sqlalchemy = "^1.4.44"
pydantic = {extras = ["email"], version = "^1.10.2"}
alembic = "^1.8.1"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
